requirements.txt
asgiref==3.3.4
dj-database-url==1.0.0
Django==3.2.6
django-appconf==1.0.5
django-compressor==2.4
django-filter==2.4.0
django-heroku==0.3.1
django-libsass==0.7
djangorestframework==3.12.4
gunicorn==20.1.0
libsass==0.21.0
Pillow==8.2.0
psycopg2==2.9.3
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.1
whitenoise==6.2.0
